requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
django-environ==0.12.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
django-filter==23.5
psycopg2-binary==2.9.9
